+++ server/visitor.go
@@ -159,8 +159,9 @@ func (v *visitor) contextNoLock() log.Context {
 		fields["user_id"] = v.user.ID
 		fields["user_name"] = v.user.Name
 		if v.user.Tier != nil {
-			fields["tier_id"] = v.user.Tier.ID
-			fields["tier_name"] = v.user.Tier.Name
+			for field, value := range v.user.Tier.Context() {
+				fields[field] = value
+			}
 		}
 		if v.user.Billing.StripeCustomerID != "" {
 			fields["stripe_customer_id"] = v.user.Billing.StripeCustomerID
@@ -331,7 +332,7 @@ func (v *visitor) SetUser(u *user.User) {
 	shouldResetLimiters := v.user.TierID() != u.TierID() // TierID works with nil receiver
 	v.user = u
 	if shouldResetLimiters {
-		v.resetLimitersNoLock(0, 0, true)
+		v.resetLimitersNoLock(u.Stats.Messages, u.Stats.Emails, true)
 	}
 }
